Legendre Polynomials

Tutorial

The Legendre Polynomials are a set of univariate polynomials defined on the interval $[-1, 1]$. Let
$ϕ_j(x) : [-1, 1] → \mathbb{R}$ denote the order $j \in \mathbb{N}_{≧0}$ Legendre Polynomial. The Legendre Polynomials are defined by the property

\[∫_{-1}^1 ϕ_i(x) ϕ_j(x) dx = 0, \quad \text{for all } i ≠ j, \quad ϕ_0(x) = 1.\]

This property is referred to orthogonality. The theoretical benefits of orthogonality are beyond the scope of this documentation. For more detailed information about the theoretical properties of the Legendre Polynomials, see here. It turns out that Legendre polynomials can be defined by the second order recursive equation

\[(n + 1) ϕ_{n + 1}(x) = (2 n + 1) x ϕ_n(x) - n ϕ_{n - 1}(x).\]

We prodive two simple methods for evaluating Legendre Polynomials. The simplest approach is by using RecursivePartition.legendre_poly.

using RecursivePartition

legendre_poly(5, 0.5) # 5th order Legendre Polynomial evaluated as x = 0.5

# output

0.08984375

However, often the user wants to evaluate the Legendre Polynomials for a range of values e.g. $j = 0, \ldots, 5$. In this case, I recommend using the recursive definintion (RecursivePartition.legendre_next) directly.

lp = Vector{Float64}(undef, 6)
lp[1:2] = [1.0, 0.5]  # Must ϕ_0(x) and ϕ_1(x) as a start.

for j in 3:6 # Remember we are starting from order 0.
  lp[j] = legendre_next(j - 1, 0.5, lp[j - 1], lp[j - 2])
end

lp

# output

6-element Array{Float64,1}:
  1.0
  0.5
 -0.125
 -0.4375
 -0.2890625
  0.08984375

Functions

RecursivePartition.legendre_next — Method

legendre_next(ord, x, d1, d0)

Recursive definition of the Legendre polynomials.

The Legendre Polynomials can be expressed as second order recursive equation, thus computing an order $n > 2$ Legendre Polynomial is possible if we know the corresponding order $n - 1$ and $n - 2$ evaluations. These are supplied in the arguments d1 and d0 (note that the order 1 and 0 evaluations are always equal to x and 1.0 respectively).

See also: legendre_poly

Examples

ord = 2
x = 1.0
d1 = 1.0
d0 = 1.0

legendre_next(ord, x, d1, d0)

# output

1.0

RecursivePartition.legendre_poly — Method

legendre_poly(ord, x)

Return the Legendre polynomial of specified order (ord) evaluated at x.

See also: legendre_next

Examples

legendre_poly(1, 1.0)

# output

1.0