Regression

The objective of regression is to model the relationship between a set of predictor and response variables. Once trained, the regression model can be used to predict the response variable, given a new predictor variable. RecursivePartition implements a variety of regression models, however its most notable feature is the PartitionModel type. This model is fitted by partitioning the space of predictor variables (see Recursive Partitioning for a visualization of this process) and fitting a flexible models within each subregion. One can think of this as a divide and conquer approach to learning complex relationships between predictors and responses.

Bayesian Linear Models

Partitioned Models

RecursivePartition.BLMHyper — Type

BLMHyper(dim::Int64; shape::Float64=0.001, scale::Float64=0.001)
BLMHyper(coeff::Vector{Float64}, cov::Matrix{Float64},
    covinv::Matrix{Floa64}, shape::Float64, scale::Float64)

Return container for the hyperparamters associated with a Bayesian linear model.

If only the dimension/shape/scale parameters are supplied, the default constructor uses the identity matrix as the covariance matrix, and a vector of zeros for the the prior mean of the model coefficients.

See also: BayesLinearModel

RecursivePartition.BayesLinearModel — Type

BayesLinearModel(X::Matrix{Float64}, y::Vector{Float64}; shape::Float64=0.001,
        scale::Float64=0.001)
BayesLinearModel(X::Matrix{Float64}, y::Vector{Float64}, prior::BLMHyper)

Construct BayesLinearModel object. One must only supply the dimensiojn of the linear model, and an optional set of prior hyper parameters. Once the model is constructed, fit! is used to train it.

This object implements the well known Bayesian Linear Model with Gaussian responses and unknown variance. A Gaussian/inverse-gamma prior is placed on the model coefficients/variance respectively.

See also: BLMHyper, fit!

Examples

coeff = [1.0, 2.0, 3.0]
f(x) = coeff[1] + coeff[2:3]' * x # Truly linear model
SD = 1.0
X, y = gendat(50000, SD, f, 2)
model = BayesLinearModel(X, y)

get_coeffpost(model)

# output

3-element Array{Float64,1}:
 0.997262172068708
 1.9957265287418318
 3.0063668906195202

RecursivePartition.LinearModel — Type

Abstract type encompassing different linear models.

See also: BayesLinearModel, PolyBLM

RecursivePartition.PartitionHyper — Type

PartitionHyper(shape::Float64, scale::Float64,
        logdet::Union{Nothing, Vector{Float64}})

Container for hyperparamers which are shared across all subregion models.

Shape and scale describe the (prior/posterior) distribution of the unknown noise across the entire space. Each local subregion model will have its own independent shape/scale parameters, which are aggregated with the other subregions to compute the shared hyperparameters.

RecursivePartition.PartitionModel — Type

PartitionModel{T <: LinearModel}

Struct containing a recursive partition and LinearModel's relating to each subset.

RecursivePartition.PolyBLM — Type

PolyBLM(X, y, degmax, bounds; maxparam=200, shape=0.001, scale0.001,
    priorgen=identity_hyper)

Construct a linear model using features derrived from the Polynomial Chaos Basis.

A bound on the maximum number of model parameters, maxparam, is specified by the user. If the number of parameters exceeds this bound, the LARS algorithm is used to choose the "best" set of parameters satisfying the bound. Consequently, a PolyBLM cannot be constructed without data (unlike BayesLinearModel).

Since we do not know which/how many parameters will be included in the model (we only that the number of parameters is bounded by maxparam), it is not possible to supply an object of type BLMHyper as a prior distribution. Instead, the argument priorgen accepts a function which specifies how the prior distribution should be generated. priorgen must have the signature

priorgen(indices::Vector{MVPIndex}, shape::Float64, scale::Float64)

and return an object of type BLMHyper

See also: BayesLinearModel, BLMHyper, MVPIndex

RecursivePartition.auto_partition_blm — Method

autopartitionblm(X::Matrix{Float64}, y::Vector{Float64}, bounds::Union{Matrix{Float64}, Vector{Float64}}; mindat=nothing, Kmax=200, shape=0.001, scale=0.001, verbose=false) autopartitionblm(X::Matrix{Float64}, y::Vector{Float64}, bounds::Union{Matrix{Float64}, Vector{Float64}}, prior::BLMHyper; mindat=nothing, Kmax=200, verbose=false)

Adaptively contruct a partition and fit a PartitionModel, using BayesLinearModel models in each subregion.

RecursivePartition.auto_partition_polyblm — Method

auto_partition_polyblm(X::Matrix{Float64}, y::Vector{Float64},
        bounds::Union{Matrix{Float64}, Vector{Float64}}; degmax::Int64=3,
        maxparam::Int64=200, priorgen::Function=identity_hyper,
        shape::Float64=0.001, scale::Float64=0.001, mindat=nothing, Kmax=200,
        verbose=false)

Adaptively contruct a partition and fit a PartitionModel, using PolyBLM models in each subregion.

RecursivePartition.fit! — Method

fit!(mod::BayesLinearModel, X::Matrix{Float64}, y::Vector{Float64})

Update a Bayesian Linear Model object with data matrix and response variables.

See also: BayesLinearModel

Examples

coeff = [1.0, 2.0, 3.0]
f(x) = coeff[1] + coeff[2:3]' * x # Truly linear model
SD = 1.0
X1, y1 = gendat(25000, SD, f, 2)
X2, y2 = gendat(25000, SD, f, 2)
model = BayesLinearModel(X1, y1) # Fit initial model
fit!(model, X2, y2) # Update model with more data

get_coeffpost(model)

# output

3-element Array{Float64,1}:
 0.996650221433608
 1.9907052757630952
 2.9924622031151826

RecursivePartition.logevidence — Method

logevidence(mod::LinearModel)

Compute the Bayesian Model Evidence/Marginal Likelihood.

RecursivePartition.partition_blm — Method

partitionblm(X::Matrix{Float64}, y::Vector{Float64}, P::Matvec; shape::Float64=0.001, scale::Float64=0.001) partitionblm(X::Matrix{Float64}, y::Vector{Float64}, P::Matvec, prior::BLMHyper)

Fit recursively partitioned [BayesLinearModel]'s with a fixed (and prespecified) partition.

RecursivePartition.partition_polyblm — Method

partition_polyblm(X::Matrix{Float64}, y::Vector{Float64},
        P::Matvec; degmax::Int64=3, maxparam::Int64=200,
        priorgen::Function=identity_hyper, shape::Float64=0.001,
        scale::Float64=0.001)

Fit recursively partitioned [PolyBLM] models with a fixed (and prespecified) partition.

RecursivePartition.predfun — Method

predfun(mod::LinearModel)

Return a function which predicts response variables given an input vector x.

RecursivePartition.predict — Method

predict(mod::BayesLinearModel, X::Matrix{Float64})

Predict responses based on data matrix X.

```

RecursivePartition.split_subset — Method

split_subset(mod::PartitionModel{T}, stored::Vector{SubsetMem{T}},
        k::Int64, dim::Int64, loc::Float64, mindat::Int64,
        args::ModArgs) where T <: LinearModel

Return a PartitionModel where the k'th subset has been divided in a specficied dimension/location.